Add files via upload
In this project, we are talking about an automated system that includes whole watering system. 
An automated watering system is such a system that starts watering to plants by measuring soil moisture through soil moisture sensor. In this project we will have some extra benefits. We don’t need to control this watering, because the device will make this automatically when it will need it.